+++ pairecservice-20221213/src/Models/ListTrafficControlTasksResponseBody/trafficControlTasks.php
@@ -35,6 +35,16 @@ class trafficControlTasks extends Model
     public $description;
 
     /**
+     * @var int[]
+     */
+    public $effectiveSceneIds;
+
+    /**
+     * @var int[]
+     */
+    public $effectiveSceneNames;
+
+    /**
      * @var string
      */
     public $endTime;
@@ -130,6 +140,11 @@ class trafficControlTasks extends Model
     public $serviceId;
 
     /**
+     * @var string[]
+     */
+    public $serviceIds;
+
+    /**
      * @var string
      */
     public $startTime;
@@ -184,6 +199,8 @@ class trafficControlTasks extends Model
         'controlLogic' => 'ControlLogic',
         'controlType' => 'ControlType',
         'description' => 'Description',
+        'effectiveSceneIds' => 'EffectiveSceneIds',
+        'effectiveSceneNames' => 'EffectiveSceneNames',
         'endTime' => 'EndTime',
         'everPublished' => 'EverPublished',
         'executionTime' => 'ExecutionTime',
@@ -203,6 +220,7 @@ class trafficControlTasks extends Model
         'sceneId' => 'SceneId',
         'sceneName' => 'SceneName',
         'serviceId' => 'ServiceId',
+        'serviceIds' => 'ServiceIds',
         'startTime' => 'StartTime',
         'statisBahaviorConditionExpress' => 'StatisBahaviorConditionExpress',
         'statisBehaviorConditionArray' => 'StatisBehaviorConditionArray',
@@ -217,6 +235,15 @@ class trafficControlTasks extends Model
 
     public function validate()
     {
+        if (\is_array($this->effectiveSceneIds)) {
+            Model::validateArray($this->effectiveSceneIds);
+        }
+        if (\is_array($this->effectiveSceneNames)) {
+            Model::validateArray($this->effectiveSceneNames);
+        }
+        if (\is_array($this->serviceIds)) {
+            Model::validateArray($this->serviceIds);
+        }
         if (\is_array($this->trafficControlTargets)) {
             Model::validateArray($this->trafficControlTargets);
         }
@@ -244,6 +271,28 @@ class trafficControlTasks extends Model
 
         if (null !== $this->description) {
             $res['Description'] = $this->description;
+        }
+
+        if (null !== $this->effectiveSceneIds) {
+            if (\is_array($this->effectiveSceneIds)) {
+                $res['EffectiveSceneIds'] = [];
+                $n1 = 0;
+                foreach ($this->effectiveSceneIds as $item1) {
+                    $res['EffectiveSceneIds'][$n1] = $item1;
+                    ++$n1;
+                }
+            }
+        }
+
+        if (null !== $this->effectiveSceneNames) {
+            if (\is_array($this->effectiveSceneNames)) {
+                $res['EffectiveSceneNames'] = [];
+                $n1 = 0;
+                foreach ($this->effectiveSceneNames as $item1) {
+                    $res['EffectiveSceneNames'][$n1] = $item1;
+                    ++$n1;
+                }
+            }
         }
 
         if (null !== $this->endTime) {
@@ -320,6 +369,17 @@ class trafficControlTasks extends Model
 
         if (null !== $this->serviceId) {
             $res['ServiceId'] = $this->serviceId;
+        }
+
+        if (null !== $this->serviceIds) {
+            if (\is_array($this->serviceIds)) {
+                $res['ServiceIds'] = [];
+                $n1 = 0;
+                foreach ($this->serviceIds as $item1) {
+                    $res['ServiceIds'][$n1] = $item1;
+                    ++$n1;
+                }
+            }
         }
 
         if (null !== $this->startTime) {
@@ -400,6 +460,28 @@ class trafficControlTasks extends Model
             $model->description = $map['Description'];
         }
 
+        if (isset($map['EffectiveSceneIds'])) {
+            if (!empty($map['EffectiveSceneIds'])) {
+                $model->effectiveSceneIds = [];
+                $n1 = 0;
+                foreach ($map['EffectiveSceneIds'] as $item1) {
+                    $model->effectiveSceneIds[$n1] = $item1;
+                    ++$n1;
+                }
+            }
+        }
+
+        if (isset($map['EffectiveSceneNames'])) {
+            if (!empty($map['EffectiveSceneNames'])) {
+                $model->effectiveSceneNames = [];
+                $n1 = 0;
+                foreach ($map['EffectiveSceneNames'] as $item1) {
+                    $model->effectiveSceneNames[$n1] = $item1;
+                    ++$n1;
+                }
+            }
+        }
+
         if (isset($map['EndTime'])) {
             $model->endTime = $map['EndTime'];
         }
@@ -474,6 +556,17 @@ class trafficControlTasks extends Model
 
         if (isset($map['ServiceId'])) {
             $model->serviceId = $map['ServiceId'];
+        }
+
+        if (isset($map['ServiceIds'])) {
+            if (!empty($map['ServiceIds'])) {
+                $model->serviceIds = [];
+                $n1 = 0;
+                foreach ($map['ServiceIds'] as $item1) {
+                    $model->serviceIds[$n1] = $item1;
+                    ++$n1;
+                }
+            }
         }
 
         if (isset($map['StartTime'])) {